refactor: move evaluateOptions initialization to evalCommand (#1674)
