+++ src/main.ts
@@ -21,7 +21,7 @@ import { runDbMigrations } from './migrate';
 import { generateRedteamCommand } from './redteam/commands/generate';
 import { initCommand as redteamInitCommand } from './redteam/commands/init';
 import { pluginsCommand as redteamPluginsCommand } from './redteam/commands/plugins';
-import { type EvaluateOptions, type UnifiedConfig } from './types';
+import { type UnifiedConfig } from './types';
 import { checkForUpdates } from './updates';
 
 export async function loadDefaultConfig(): Promise<{
@@ -51,21 +51,14 @@ async function main() {
   await checkForUpdates();
   await runDbMigrations();
 
-  const program = new Command();
-
   const { defaultConfig, defaultConfigPath } = await loadDefaultConfig();
 
-  const evaluateOptions: EvaluateOptions = {};
-  if (defaultConfig.evaluateOptions) {
-    evaluateOptions.generateSuggestions = defaultConfig.evaluateOptions.generateSuggestions;
-    evaluateOptions.maxConcurrency = defaultConfig.evaluateOptions.maxConcurrency;
-    evaluateOptions.showProgressBar = defaultConfig.evaluateOptions.showProgressBar;
-  }
+  const program = new Command();
 
   cacheCommand(program);
   configCommand(program);
   deleteCommand(program);
-  evalCommand(program, defaultConfig, defaultConfigPath, evaluateOptions);
+  evalCommand(program, defaultConfig, defaultConfigPath);
   exportCommand(program);
   feedbackCommand(program);
   importCommand(program);
